+++ src/entities/UserAvatar/UserAvatar.tsx
@@ -8,14 +8,15 @@ type Props = {
 }
 
 const UserAvatar = ({ avatar, userName }: Props) => {
-  return avatar ? (
+  return (
     <div className="flex gap-[12px] items-center py-3">
-      <Image alt="avatar" className="rounded-full" height={36} src={avatar} width={36} />
-      <p>{userName}</p>
-    </div>
-  ) : (
-    <div className="flex gap-[12px] items-center py-3">
-      <Image alt="avatar" className="rounded-full" height={36} src="/avatar.png" width={36} />
+      <Image
+        alt="avatar"
+        className="rounded-full"
+        height={36}
+        src={avatar ? avatar : '/avatar.png'}
+        width={36}
+      />
       <p>{userName}</p>
     </div>
   )
